% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddID.r
\name{add_id}
\alias{add_id}
\title{Add a Custom ID Column to Data Frame}
\usage{
add_id(data, prefix, pad_number = 10, name = "ID")
}
\arguments{
\item{data}{\code{data.frame} or \code{data.table}}

\item{prefix}{Chacter string that appears befoer the number}

\item{pad_number}{Integer indicating the total digits (not inlcluding prefix)}

\item{name}{Character string for name of the variable}
}
\description{
\code{add_id} adds a sequestional ID column with a custom prefix.
}
\note{
For \code{data.table} load package \code{dtplyr}
}
\examples{
foo <- data.frame(a = c(1:10), b = c(a:z))
foo <- AddID(foo, "test")

foo2 <- data.frame(a = c(1:10), b = c(a:z))
foo2 <- AddID(foo2, "PRE", pad_number = 3, name = "Tag")

}

