% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddID.r
\name{AddID}
\alias{AddID}
\alias{AddID.data.table}
\alias{AddID.default}
\title{Add a Custom ID Column to Data Frames}
\usage{
AddID(data, prefix, pad_number = 10, name = "ID")

\method{AddID}{default}(data, prefix, pad_number = 10, name = "ID")

\method{AddID}{data.table}(data, prefix, pad_number = 10, name = "ID")
}
\arguments{
\item{data}{a data frame}

\item{prefix}{a chacter string}

\item{pad_number}{an integer indicating the total digits (inlcluding padding)}

\item{name}{a character string that appears before the number}
}
\value{
For default a data frame is returned. If a \code{\link[data.table]{data.table}} then
 a data table will be returned \strong{invisibly}
}
\description{
\code{AddID} adds a sequestional ID column with a custom prefix.
This is a generic function that has two methods (defulat and data.table).
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: data frame must be assigned

\item \code{data.table}: data table returned invisibly
}}
\examples{
foo <- data.frame(a = c(1:10), b = c(a:z))
foo <- AddID(foo, "test")

foo2 <- data.frame(a = c(1:10), b = c(a:z))
foo2 <- AddID(foo2, "PRE", .pad_number = 3, .name = "Tag")

foo3 <- data.table(a = c(1:10), b = c(a:z))
AddID(foo3, "ROW", .name = "Row") #Returned Invisibily


}

