% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cap_and_Floor.R
\name{Cap and Floor}
\alias{Cap and Floor}
\title{Functions to Cap, Floor and calulate Max values of a \code{data.frame}}
\description{
These function are for manipulating a data.frame for modeling
}
\section{Cap}{

\code{Cap} the maximum value for a \code{data.frame} column
\itemize{
   \item \bold{data} {a \code{data.frame} or \code{data.table}}
   \item \bold{var} {the name of the variable to be capped}
   \item \bold{cap} {a number that is will be the max for the column}
}
\bold{Returns} an object of the same class as \code(data)
}

\section{Floor}{

\code{Floor} sets the minimum value for a \code{data.frame} column
\itemize{
   \item \bold{data} {a \code{data.frame} or \code{data.table}}
   \item \bold{var} {the name of the variable to be floored}
   \item \bold{floor} {a number that is will be the min for the column}
}
\bold{Returns} an object of the same class as \code(data)
}

\section{Max}{

\code{Max} calculates the max value less than than the \code{value}
\itemize{
   \item \bold{data.var} {a numeric vector (usually from a \code{data.frame})}
   \item \bold{value} {the maximum number to limit the vector to (i.e. \code{max(x < value)})}
   \item \bold{na_rm} {logical, to indicate if NAs should be removed}
}
\bold{Returns:} number of max value
}

