% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cap_and_Floor.R
\name{Cap_and_Floor}
\alias{Cap_and_Floor}
\alias{Max}
\alias{cap}
\alias{cap_all}
\alias{floor}
\alias{floor_all}
\title{Functions to Cap, Floor and calulate Max values}
\usage{
cap(data, cap)

cap_all(data, cap)

floor(data, floor)

floor_all(data, floor)

Max(data.var, value, na_rm = TRUE)
}
\arguments{
\item{data}{Numeric vector}

\item{cap}{Number of max value}

\item{floor}{Number of min value}
}
\description{
These function are for manipulating a data.frame for modeling
}
\section{Max}{

\code{Max} calculates the max value less than than the \code{value}
\itemize{
   \item \bold{data.var} {a numeric vector (usually from a \code{data.frame})}
   \item \bold{value} {the maximum number to limit the vector to (i.e. \code{max(x < value)})}
   \item \bold{na_rm} {logical, to indicate if NAs should be removed}
}
\bold{Returns:} number of max value
}

