% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackage.R
\name{checkPackage}
\alias{checkPackage}
\title{checkPackage function}
\usage{
checkPackage(pkg)
}
\arguments{
\item{pkg}{Name of the package to check}
}
\description{
Check if a package is installed.
If not, then install it.
If an update exists, then update.
}
\examples{
checkPackage("data.table")
}
\keyword{install.packages}
\keyword{library}

