% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cdp_plans.R
\name{clean_cdp_plans}
\alias{clean_cdp_plans}
\title{Map and Clean CDP Plans}
\usage{
clean_cdp_plans(
  data,
  cdp_map,
  plan_cols = c("Academic Plan Nm - First Plan", "Academic Plan Nm - Second Plan"),
  map_key = "Academic Plan Nm - First Plan",
  map_value = "Academic Plan Nm - NonCDP"
)
}
\arguments{
\item{data}{The data frame with CDP academic plans}

\item{cdp_map}{The mapping table (CDP to non-CDP plan names)}

\item{plan_cols}{A character vector of the column names to be cleaned (e.g., c("Plan 1", "Plan 2"))}

\item{map_key}{The column in cdp_map that matches the academic plan names}

\item{map_value}{The column in cdp_map containing the "NonCDP" replacement text}
}
\value{
The original data frame with the academic plan column text values updated
}
\description{
Map and Clean CDP Plans
}
\seealso{
Other institutional data prep: 
\code{\link{add_ac_masters_en_route_phd}()},
\code{\link{add_program_type_classification}()}
}
\concept{institutional data prep}
