% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanEmail.r
\name{clean_email}
\alias{clean_email}
\title{Removes Invalid Characters from Email Addresses}
\usage{
clean_email(var)
}
\arguments{
\item{var}{\code{character} vector of email addresses}
}
\value{
a character vector of the same length
}
\description{
\code{clean_email} removes invalid characters and phrases
 such as "N/A", "NOEMAIL", and " "
}

