% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_misc.r
\name{clean_name}
\alias{clean_name}
\title{Removes Bad and Unwated Characters and Wrods from Names}
\usage{
clean_name(data, args_list)
}
\description{
Can remove a predetermined set of bad words and/or a custom list of words or
 regular expresions
}

