% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanEmail.r
\name{CleanEmail}
\alias{CleanEmail}
\title{Removes Invalid Characters from Email Addresses}
\usage{
CleanEmail(var)
}
\arguments{
\item{var}{a character vector of email addresses}
}
\value{
a character vector of the same length
}
\description{
\code{CleanEmail} removes invalid characters and phrases
 such as "N/A", "NOEMAIL", and " "
}

