% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_left_join.R
\name{coalesce_left_join}
\alias{coalesce_left_join}
\title{Coalescing Left Join}
\usage{
coalesce_left_join(x, y, by = NULL, suffix = c(".x", ".y"), ...)
}
\arguments{
\item{x, y}{Tables to join.}

\item{by}{A character vector of variables to join by.}

\item{suffix}{Suffixes to apply to overlapping columns.}
}
\description{
Performs a left join and merges overlapping columns, filling NAs in the
left table with values from the right table.
}
\examples{
df1 <- tibble::tibble(ID = 1:2, val = c("A", NA))
df2 <- tibble::tibble(ID = 1:2, val = c("Z", "B"))

# Standard join creates val.x and val.y
# Coalesce join fills the NA in df1 with the value from df2
coalesce_left_join(df1, df2, by = "ID")

}
