\name{colon}
\docType{data}
\alias{colon}
\title{Colon cancer data}
\seealso{
  \code{\link{colon.example}}.
}
\description{The colorectal cancer data set, originally published in Alon et al. (1999) contains the expressions of 2000 genes collected from 62 tissues. The gene expression levels are reported in the descending order of minimal intensity.}
\usage{colon}
\format{A data frame with 2001 columns (variables) and 62 rows (samples). The first column represents the phenotype of the samples with "T" indicating tumor tissue and "N" indicating a normal tissue. The actual data containing the microarray expression values of 2000 genes across 62 patients is presented in the remaining columns.}
\source{http://genomics-pubs.princeton.edu/oncology/}
\references{
        \emph{Broad patterns of gene expression revealed by clustering of tumor and normal colon tissues probed by oligonucleotide arrays}; U. Alon, N. Barkai, D. A. Notterman, K. Gish, S. Ybarra, D. Mack, and A. J. Levine, Proc. Natl. Acad. Sci. , 96(12), pp. 6745-6750
}
\keyword{datasets}
