\name{colon.example}
\alias{colon.example}
\title{Example - Colon cancer data}
\description{
    Run the prepivoting root based test (PREPR) on the Colon cancer gene expression data from Alon et al., 1999.
}
\usage{
    colon.example()
}
\value{
    The pre-pivoting test statistic, p-value and decision as obtained from \link{prepr.test}
}
\seealso{
  \code{\link{colon}}.
}

\keyword{example}
