% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPalette.R
\name{colorPalette}
\alias{colorPalette}
\title{colorPalette function}
\usage{
colorPalette()
}
\description{
Defines a custom color palette.
Returns a vector of RGB values.
}
\examples{
colorPalette()
}
\keyword{rgb}

