% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Assertions.R
\name{Custom_Assertions}
\alias{Custom_Assertions}
\alias{assert_number_greater}
\alias{is_integer}
\title{Custom Assertions}
\usage{
is_integer(x)

assert_number_greater(x, n)
}
\arguments{
\item{x}{object to be checked}

\item{n}{\code{integer}}
}
\description{
Assertions for \code{\link{testthat}}. Not exported and not available.
}
\section{Assertions}{

\itemize{
 \item{\code{is_integer} checks if a number is an integer
   (\code{is_integer(5)} returns \code{TRUE})
 }
 \item{\code{assert_number_greater} checks if \code{x} is great than \code{n}}
}
}

