% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{data_type}
\alias{data_type}
\title{data_type
Identify features of different data types.}
\usage{
data_type(df)
}
\arguments{
\item{df}{data.frame
Original feature dataframe containing one column for each feature.}
}
\value{
list of data.frame ($num, $cat)
Stores the categorical and numerical columns separately as two dataframes in one list.
The first element in the list will contain categorical dataframe and the second will contain numerical dataframe.
}
\description{
data_type
Identify features of different data types.
}
\examples{
my_data <- data.frame(fruits = c('apple', 'banana', 'pear'), counts = c(1, 2, 3), price = c(0, 1, 2))
data_type(my_data)$num
data_type(my_data)$cat
}
