% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Substrate.R
\name{excel_functions}
\alias{excel_functions}
\alias{ifna}
\alias{left}
\alias{right}
\title{Functions that are similar to excel functions}
\usage{
right(vec1, chars)

left(vec1, chars)

ifna(var1, var2)
}
\arguments{
\item{vec1}{Charater vector}

\item{chars}{Number of characters to substrate}

\item{var1}{Vector to check if \code{NA}}

\item{var2}{Value to replace \code{NA}}
}
\description{
\code{right} and \code{left} act like Right and Left excel functions
\code{ifna} replaces NAs with the provided value
}

