% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSQL.R
\name{file_location}
\alias{file_location}
\title{file_location}
\usage{
file_location(f.file)
}
\arguments{
\item{f.file}{character string of folders and file name}
}
\description{
\code{file_location} create a character string using the current working directory
 and provided file location.
}
\note{
\code{f.file} must begin with a '\' and cannot use '/'
}

