% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FixNames.R
\name{FixName}
\alias{FixName}
\title{Cleans and parses combined name fields}
\usage{
FixName(.data, .name_field)
}
\arguments{
\item{.data}{\code{\link[data.table]{data.table}}}
}
\value{
\code{data.table} invisibly
}
\description{
\code{FixName} prosses a raw name column, parses the name
  into, at most, 5 words, and assigns words to Prefix, First,
  Middle, Last, and Suffix.
 Only for \code{\link[data.table]{data.table}} and all results returned invisibly
}

