% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{form_parts}
\alias{form_parts}
\title{Parse Formula for Variable Specification}
\usage{
form_parts(f, data)
}
\arguments{
\item{f}{A formula specifying how variables should be treated.}

\item{data}{A dataset containing variables referenced by the formula.}
}
\value{
A nested list containing information \link{\code{Transformer}} objects
need for processing fit & transform methods.
}
