% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboard_varinfo.R
\name{get_dashboard_varinfo}
\alias{get_dashboard_varinfo}
\title{Slim Varinfo for Export for Dashboard Display}
\usage{
get_dashboard_varinfo(df, display_col = "DASH_DISPLAY")
}
\arguments{
\item{df}{The varinfo data frame.}

\item{display_col}{String. The name of the boolean flag column to filter on.
Defaults to "DASH_DISPLAY".}
}
\value{
A filtered and column-trimmed tibble.
}
\description{
Filters the varinfo to only include variables marked for a specific
display and selects a specific subset of columns required by the dashboard.
}
\examples{
# Export for Program Dashboard
dash_export <- get_dashboard_varinfo(varinfo, "DASH_DISPLAY")

# Export for General Dashboards
general_viz_export <- get_dashboard_varinfo(varinfo, "GENERAL_VIZ_DISPLAY")
}
\seealso{
Other varinfo prep functions: 
\code{\link{add_catalog_display_flag}()},
\code{\link{add_dash_display_flag}()},
\code{\link{add_display_flag}()},
\code{\link{add_label_metadata}()},
\code{\link{check_and_export_unmatched}()},
\code{\link{get_catalog_varinfo}()},
\code{\link{join_varinfo}()},
\code{\link{load_survey_column_info}()},
\code{\link{sort_varinfo}()},
\code{\link{sync_column_schema}()},
\code{\link{update_survey_admin_summaries}()}
}
\concept{varinfo prep functions}
