% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMetadata.R
\name{makeMetadata}
\alias{makeMetadata}
\title{makeMetadata function}
\usage{
makeMetadata(D, note = NULL)
}
\arguments{
\item{D}{data.frame object to make metadata for}

\item{note}{character object containing annotation}
}
\description{
Make metadata associated with a data.frame object.
}
\examples{
makeMetadata(iris)
makeMetadata(iris, note="Edgar Anderson's Iris Data")
}
\keyword{data.frame}

