% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_name.R
\name{name_word_functions}
\alias{five_words}
\alias{four_words}
\alias{name_word_functions}
\alias{three_words}
\alias{two_words}
\title{Functions that process name words}
\usage{
two_words(word1, word2)

three_words(word1, word2, word3)

four_words(word1, word2, word3, word4)

five_words(word1, word2, word3, word4, word5)
}
\value{
\code{list} with name components filled. Missing components are filled with
 \code{NA_caracter}
}
\description{
Function for specific word totals (i.e. 2, 3, 4, or 5 words) that
 process them into name components. \strong{Not exported}.
}

