% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnehotEncoder.R
\name{prep_onehot}
\alias{prep_onehot}
\title{Onehot Encoder}
\usage{
prep_onehot(..., keep = FALSE, cols = "factor")
}
\arguments{
\item{...}{Named arguments where the name maps to variables in the dataset and the argument is
a character vector of levels to create indicator columns}

\item{keep}{keep original factors levels after onehot encoding}

\item{cols}{character vector of column names to onehot encode. Defaults to all factor columns.}
}
\description{
Encode factor variables to indicator columns
}
\examples{
data(mtcars)
mtcars$cyl <- factor(mtcars$cyl)
prep <- prep_onehot()
prep$fit_transfrom(mtcars)
}
