\name{prepair-package}
\alias{prepair-package}
\alias{prepair}
\docType{package}
\title{\packageTitle{prepair}}
\description{\packageDescription{prepair}}
\section{Package Content}{\packageIndices{prepair}}
\author{\packageAuthor{prepair}}
\section{Maintainer}{\packageMaintainer{prepair}}
\keyword{package}
