% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepair-package.R
\docType{package}
\name{prepair}
\alias{prepair}
\alias{prepair-package}
\title{A package for repairing broken GIS polygons}
\description{
It's based on the prepair Cpp library, more info at \url{https://github.com/tudelft3d/prepair}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/dickoa/prepair}
  \item Report bugs at \url{https://gitlab.com/dickoa/prepair/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com}

}
