% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{prepr-package}
\alias{prepr}
\alias{prepr-package}
\title{A package for automatically repairing broken polygons}
\description{
This package contains functions to repair single polygons according to the
international standards ISO 19107 using a constrained triangulation approach.
It is based on the prepair C++ tool.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/prioritizr/prepr}
  \item Report bugs at \url{https://github.com/prioritizr/prepr/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey O Hanson \email{jeffrey.hanson@uqconnect.edu.au} (\href{https://orcid.org/0000-0002-4716-6134}{ORCID})

Authors:
\itemize{
  \item Ahmadou Dicko \email{mail@ahmadoudicko.com}
}

Other contributors:
\itemize{
  \item Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID}) [contributor]
  \item Ken Arroyo Ohori \email{g.a.k.arroyoohori@tudelft.nl} (https://github.com/tudelft3d/prepair) [contributor, copyright holder]
  \item Hugo Ledoux \email{h.ledoux@tudelft.nl} (https://github.com/tudelft3d/prepair) [contributor, copyright holder]
  \item Martijn Meijers \email{b.m.meijers@tudelft.nl} (https://github.com/tudelft3d/prepair) [contributor, copyright holder]
}

}
