% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_bundle.R
\name{print.qc_bundle}
\alias{print.qc_bundle}
\title{Print method for qc_bundle}
\usage{
\method{print}{qc_bundle}(x, ...)
}
\arguments{
\item{x}{qc_bundle}

\item{...}{ignored}
}
\description{
Print method for qc_bundle
}
