% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_bundle.R
\name{print.summary.qc_bundle}
\alias{print.summary.qc_bundle}
\title{Print method for summary.qc_bundle}
\usage{
\method{print}{summary.qc_bundle}(x, ...)
}
\arguments{
\item{x}{summary.qc_bundle}

\item{...}{ignored}
}
\description{
Print method for summary.qc_bundle
}
