% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_bundle.R
\name{qc_enable_dummy}
\alias{qc_enable_dummy}
\title{Dummy QC enabling step for testing logs (internal)}
\usage{
qc_enable_dummy(bundle, unit = "day", window = 60L)
}
\description{
Dummy QC enabling step for testing logs (internal)
}
\keyword{internal}
