% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_bundle.R
\name{qc_init}
\alias{qc_init}
\title{Initialize a QC bundle}
\usage{
qc_init(data, meta = list(), validate = FALSE)
}
\arguments{
\item{data}{Raw data (data.frame / tibble / other).}

\item{meta}{A list of metadata describing `data`.}
}
\value{
An object of class `qc_bundle`.
}
\description{
Create a `qc_bundle` object that carries raw data and QC metadata containers.
This function performs no QC and does not modify the raw data.
}
