% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_bundle.R
\name{qc_log_init}
\alias{qc_log_init}
\title{Initialize QC audit log (internal)}
\usage{
qc_log_init()
}
\value{
A data.frame representing an append-only audit log.
}
\description{
Initialize QC audit log (internal)
}
\keyword{internal}
