% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_bundle.R
\name{qc_stage_assert_write}
\alias{qc_stage_assert_write}
\title{Assert write permission for qc_bundle fields (internal)}
\usage{
qc_stage_assert_write(bundle, stage, fields)
}
\arguments{
\item{bundle}{qc_bundle}

\item{stage}{stage name}

\item{fields}{character vector of field names that are about to be written}
}
\description{
Assert write permission for qc_bundle fields (internal)
}
\keyword{internal}
