% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadSQL.R
\name{ReadSQL}
\alias{ReadSQL}
\title{Read in a text file with .sql}
\usage{
ReadSQL(file)
}
\arguments{
\item{file}{character vector of .sql location}
}
\value{
character vector of text
}
\description{
\code{ReadSQL} reads in text files with .sql extension
 for use with SQL packages like sqldf and RODBC
}

