% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Substrate.R
\name{right}
\alias{right}
\title{Functions that are similar to excel functions}
\usage{
right(vec1, chars)
}
\arguments{
\item{vec1}{a charater vector}

\item{chars}{number of characters to substrate}
}
\description{
\code{right} and \code{left} act like Right and Left excel functions
\code{ifna} replaces NAs with the provided value
}

