% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSessionInfo.R
\name{showSessionInfo}
\alias{showSessionInfo}
\title{showSessionInfo function}
\usage{
showSessionInfo()
}
\description{
Show R session information for reproducibility purposes.
}
\examples{
.timeStart <- Sys.time()  # Defined prior to calling showSessionInfo()
showSessionInfo()
showSessionInfo()$timeEnd
}
\keyword{debugging}
\keyword{reproducibility}
\keyword{trace}

