% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_prepair.R
\name{st_prepair}
\alias{st_prepair}
\title{Automatic repair of polygon geometries}
\usage{
st_prepair(x)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}}, \code{\link[sf:sfc]{sf::st_sfc()}} or \code{\link[sf:st]{sfg}}
object (containing \code{POLYGON} or \code{MULTIPOLYGON} geometries).}
}
\value{
A \code{\link[sf:sf]{sf::st_sf()}}, \code{\link[sf:sfc]{sf::st_sfc()}} or \code{\link[sf:st]{sfg}}
object (same as the argument to \code{x}).
}
\description{
Repair polygon geometries according to the international standards ISO 19107
using a constrained triangulation approach
(van Oosterom et al. 2005; Ledoux et al. 2014)
}
\details{
The function supports two algorithms:
\itemize{
\item oddeven: an extension of the odd-even algorithm to handle polygons
containing inner rings and degeneracies;
\item setdiff: one where we follow a point set difference rule for the rings
(outer - inner).
}
}
\examples{
\dontrun{
# create an object containing a broken polygon geometry
x <- sf::st_as_sfc("POLYGON((0 0, 0 10, 10 0, 10 10, 0 0))")

# check if this polygon is indeed broken
sf::st_is_valid(x)

# repair the polygon
y <- st_prepair(x)

# check that the repaired polygon has been fixed
print(st_is_valid(y))
}

}
\references{
Ledoux H, Arroyo Ohori K, and Meijers M (2014)
A triangulation-based approach to automatically repair GIS polygons.
\emph{Computers & Geosciences} 66:121--131.

van Oosterom P, Quak W, and Tijssen T (2005)
\emph{About Invalid, Valid and Clean Polygons}
In: Developments in Spatial Data Handling. Springer, Berlin, Heidelberg
}
\seealso{
See \code{\link[sf:valid]{sf::st_make_valid()}} for another approach to repair polygon
geometries.
}
