% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{st_probustness}
\alias{st_probustness}
\title{Compute polygons robustness}
\usage{
st_probustness(x)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}, only with POLYGON or MULTIPOLYGON
are supported}
}
\description{
Compute polygons robustness
}
\examples{
library(sf)
p1 <- st_as_sfc("POLYGON((0 0, 0 10, 10 0, 10 10, 0 0))")
st_is_valid(p1)
st_probustness(p1)
st_probustness(st_prepair(p1))

}
\references{
Ledoux, H., Arroyo Ohori, K., and Meijers, M. (2014).
A triangulation-based approach to automatically repair GIS polygons.
Computers & Geosciences 66:121–131.
}
