% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardImputer.R
\docType{class}
\name{StandardImputer-class}
\alias{StandardImputer-class}
\alias{StandardImputer}
\title{Impute NA Values}
\arguments{
\item{method}{A function name as a character value}

\item{value}{A constant numeric value to impute}
}
\description{
Replace NA values with calculated values or a constant
}
\details{
If a method is passed, it will be called on each vector with
\code{na.rm=TRUE}. The return value will be stored and all NA values will
be replaced with this calculated value.

If, instead, a value is passed into the constructor it will be replace
all NA values.
}

