% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StripWhite.r
\name{strip_white}
\alias{strip_white}
\title{Strips all White Space from character string}
\usage{
strip_white(data)
}
\arguments{
\item{data}{\code{data.frame} containing character columns or \code{character vector}}
}
\value{
Object of input class
}
\description{
\code{strip_white} removes all extra spaces from before and after a character string
}

