% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StripWhite.r
\name{StripWhite}
\alias{StripWhite}
\alias{StripWhite.data.frame}
\alias{StripWhite.data.table}
\alias{StripWhite.default}
\title{Strips all White Space from character string}
\usage{
StripWhite(data)

\method{StripWhite}{default}(data)

\method{StripWhite}{data.frame}(data)

\method{StripWhite}{data.table}(data)
}
\arguments{
\item{data}{a data.frame/data.table/character vector}
}
\value{
same class as imput: data.frame/data.table/character vector
}
\description{
\code{StripWhite} removes all extra spaces from before and after a character string
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{data.frame}: 

\item \code{data.table}: Returned Invisibly
}}

