% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_bundle.R
\name{summary.qc_bundle}
\alias{summary.qc_bundle}
\title{Summary method for qc_bundle}
\usage{
\method{summary}{qc_bundle}(object, ...)
}
\arguments{
\item{object}{qc_bundle}

\item{...}{ignored}
}
\description{
Summary method for qc_bundle
}
