% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_column_schema.R
\name{sync_column_schema}
\alias{sync_column_schema}
\title{Synchronize Column Schema}
\usage{
sync_column_schema(x, y)
}
\arguments{
\item{x}{The data frame to be updated.}

\item{y}{The data frame to use as a template.}
}
\description{
Ensures that a target data frame (x) contains all columns present in a
source data frame (y), matching both names and data types.
Missing columns are initialized with NA.
}
\seealso{
Other varinfo prep functions: 
\code{\link{add_catalog_display_flag}()},
\code{\link{add_dash_display_flag}()},
\code{\link{add_label_metadata}()},
\code{\link{check_and_export_unmatched}()},
\code{\link{get_dashboard_varinfo}()},
\code{\link{join_varinfo}()},
\code{\link{load_survey_column_info}()},
\code{\link{sort_varinfo}()},
\code{\link{update_survey_admin_summaries}()}
}
\concept{varinfo prep functions}
