% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_patch_schema.R
\name{sync_patch_schema}
\alias{sync_patch_schema}
\title{Sync Schema for Upserting}
\usage{
sync_patch_schema(data, patches)
}
\arguments{
\item{data}{The target data frame (e.g., joined_draft).}

\item{patches}{The data frame containing updates (e.g., patches).}
}
\description{
Ensures that the target data frame has all columns present in the patch data frame.
Missing columns are initialized with NA.
}
