% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TitleCase.R
\name{TitleCase}
\alias{TitleCase}
\title{Puts Vector in Title Case}
\usage{
TitleCase(v)
}
\arguments{
\item{v}{a character vector}
}
\value{
a character vecto
}
\description{
\code{TitleCase} capitalizes the first letter of each word and returns the other in lowercase
}

