% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_valid_test_split.R
\name{train_valid_test_split}
\alias{train_valid_test_split}
\title{train_valid_test_split
Splits feature and target data frames into random train, validation and test subsets
The proportion of the train set relative to the input data will be valid_size * (1 - test_size)}
\usage{
train_valid_test_split(x, y, test_size = 0.25, valid_size = 0.25,
  shuffle = TRUE, random_state = NULL)
}
\arguments{
\item{x}{(data frame)
Original features from data set}

\item{y}{(data frame)
Original labels from data set}

\item{test_size}{(double, default = NULL)
Value between 0.0 and 1.0 to represent the proportion of the dataset to comprise the size of the test subset}

\item{valid_size}{float or None (default = 0.25)
Value between 0.0 and 1.0 to represent the proportion of the dataset to comprise the size of the test subset}

\item{shuffle}{(default = TRUE)
Shuffle the rows or not}

\item{random_state}{(int, default = NULL)
Seed for the random number generator}
}
\value{
list
}
\description{
train_valid_test_split
Splits feature and target data frames into random train, validation and test subsets
The proportion of the train set relative to the input data will be valid_size * (1 - test_size)
}
\examples{
x = data.frame('X1'=c(0,1,2,3,4,5,6,7),
               'X2'=c(8,9,10,11,12,13,14,15))

y = data.frame('Y'=c(0,1,2,3,4,5,6,7))


train_valid_test_split(x,y)$x_train
train_valid_test_split(x,y)$x_valid
train_valid_test_split(x,y)$x_test
train_valid_test_split(x,y)$y_train
train_valid_test_split(x,y)$y_valid
train_valid_test_split(x,y)$y_test
}
