% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_yeojohnson.R
\name{trans_yeojohnson}
\alias{trans_yeojohnson}
\title{Yeo-Johnson Transformation}
\usage{
trans_yeojohnson(x, lambda = "auto")
}
\arguments{
\item{x}{A numeric vector.}

\item{lambda}{A numeric value or "auto".}
}
\value{
A numeric vector with attribute "lambda".
}
\description{
A power transformation similar to Box-Cox but supports both positive and negative values.
Automatically estimates the optimal lambda using MLE.
}
\references{
Yeo, I.-K., & Johnson, R. A. (2000). A new family of power transformations to improve normality or symmetry. Biometrika.
}
