% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Capitalize.r
\name{Capitalize}
\alias{Capitalize}
\alias{Capitalize.data.frame}
\alias{Capitalize.data.table}
\alias{Capitalize.default}
\title{Capitalizes all Character Strings}
\usage{
Capitalize(data)

\method{Capitalize}{data.table}(data)

\method{Capitalize}{data.frame}(data)

\method{Capitalize}{default}(data)
}
\arguments{
\item{data}{a data.frame/data.table/character vector}
}
\value{
same class as was used for input
}
\description{
\code{Capitalize} is a generic function that capitalizes
 all letters in an object.
}
\details{
Current classes supported:
\itemize{
 \item \code{data.frame}
 \item \code{data.table}
 \item \code{character} vector
}
}
\section{Methods (by class)}{
\itemize{
\item \code{data.table}: Uses data.frame method (must be assigned)

\item \code{data.frame}: \code{data.frame} must be assigned

\item \code{default}: Character and Unsupported Classes
}}

