% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboard_varinfo.R
\name{get_dashboard_varinfo}
\alias{get_dashboard_varinfo}
\title{Slim Varinfo for Export for Dashboard Display}
\usage{
get_dashboard_varinfo(df)
}
\arguments{
\item{df}{The varinfo data frame.}
}
\value{
A filtered and column-trimmed tibble.
}
\description{
Filters the varinfo to only include variables marked for dashboard display
and selects a specific subset of columns required by the dashboard.
}
\seealso{
Other varinfo prep functions: 
\code{\link{add_catalog_display_flag}()},
\code{\link{add_dash_display_flag}()},
\code{\link{add_label_metadata}()},
\code{\link{check_and_export_unmatched}()},
\code{\link{join_varinfo}()},
\code{\link{load_survey_column_info}()},
\code{\link{sort_varinfo}()},
\code{\link{sync_column_schema}()},
\code{\link{update_survey_admin_summaries}()}
}
\concept{varinfo prep functions}
