% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dashboard_varinfo.R
\name{get_dashboard_varinfo}
\alias{get_dashboard_varinfo}
\title{Slim Varinfo for Export for Dashboard Display}
\usage{
get_dashboard_varinfo(df)
}
\arguments{
\item{df}{The varinfo data frame.}
}
\value{
A filtered and column-trimmed tibble.
}
\description{
Filters the varinfo to only include variables marked for dashboard display
and selects a specific subset of columns required by the dashboard.
}
