% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey_column_info.R
\name{load_survey_column_info}
\alias{load_survey_column_info}
\title{Load Qualtrics Column Metadata from Response CSV}
\usage{
load_survey_column_info(file_path, survey_admin)
}
\arguments{
\item{file_path}{Path to raw Qualtrics CSV.}

\item{survey_admin}{ID for this administration (e.g., "2025" or "Fa2024").}
}
\description{
Reads a Qualtrics CSV and extracts the variable names and question text
(usually found in the first row of data).
}
\seealso{
Other varinfo prep functions: 
\code{\link{add_catalog_display_flag}()},
\code{\link{add_dash_display_flag}()},
\code{\link{add_label_metadata}()},
\code{\link{check_and_export_unmatched}()},
\code{\link{get_catalog_varinfo}()},
\code{\link{get_dashboard_varinfo}()},
\code{\link{join_varinfo}()},
\code{\link{sort_varinfo}()},
\code{\link{sync_column_schema}()},
\code{\link{update_survey_admin_summaries}()}
}
\concept{varinfo prep functions}
