% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ParseWords.R
\name{ParseWords}
\alias{ParseWords}
\title{Parse a character column into several columns}
\usage{
ParseWords(data, total.words = 5, name_field = NULL, var.name = "word")
}
\arguments{
\item{data}{a data.frame or data.table or tbl}

\item{total.words}{numerical, total new columns}

\item{var.name}{character, new columns name prefix}
}
\value{
data.frame/data.table/tbl of the same class
}
\description{
\code{ParseWords} separates one character column (usually names)
 into several columns with sequential names.
 Originally designed for Parsing Name Fields
}

