% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseWords.R
\name{ParseWords}
\alias{ParseWords}
\title{Parse a character column into several columns}
\usage{
ParseWords(data, total_words = 5, name_field = NULL, var_name = "word")
}
\arguments{
\item{data}{\code{data.frame}}

\item{total_words}{\code{integer} total new columns}

\item{name_field}{\code{string} of column with names}

\item{var.name}{\code{string} for prefix to new columns names}
}
\value{
\code{data.frame} of the same class as \code{data}
}
\description{
\code{ParseWords} separates one character column (usually names)
 into several columns with sequential names.
 Originally designed for Parsing Name Fields.
}
\examples{
 options(stringsAsFactors = FALSE)
 x <- data.frame(text = c("Mike t Martin", "John D Rockefeller", "Daniel Tiger", "jean claude van damme"))

 # Seperate into 3 columns: name1, name2, name3
 ParseWords(x, 3, "text", "name")
}

