% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepr-package.R
\docType{package}
\name{prepr-package}
\alias{prepr}
\alias{prepr-package}
\title{A package for repairing broken GIS polygons}
\description{
It is based on the prepair Cpp library, more info at \url{https://github.com/tudelft3d/prepair}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/dickoa/prepr}
  \item Report bugs at \url{https://gitlab.com/dickoa/prepr/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com}

Authors:
\itemize{
  \item Hugo Ledoux \email{h.ledoux@tudelft.nl} (C++ original, https://github.com/tudelft3d/prepair)
  \item Ken Arroyo Ohori \email{g.a.k.arroyoohori@tudelft.nl} (C++ original, https://github.com/tudelft3d/prepair)
}

}
