% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepR-package.R
\docType{package}
\name{prepR-package}
\alias{prepR}
\alias{prepR-package}
\title{prepR: Data Prep for Grad Div IR Team}
\description{
This package contains functions for common data cleaning and preparation steps used by UC Berkeley Graduate Division's Institutional Research Team. Tasks include preparing survey response and metadata, processing academic plan information, and summarizing data for visualization and analysis.
}
\author{
\strong{Maintainer}: Madeline Arnold \email{madeline_arnold@berkeley.edu}

}
\keyword{internal}
