% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_varinfo.R
\name{sort_varinfo}
\alias{sort_varinfo}
\title{Sort Varinfo by Recency and Survey Order}
\usage{
sort_varinfo(varinfo, new_info, recency_order)
}
\arguments{
\item{new_info}{The current survey metadata from \code{load_survey_column_info}.}

\item{recency_order}{Vector of admin IDs in order, most recent first
(e.g. c("2025", "2024")).}

\item{updated_varinfo}{The varinfo table containing admin summary columns
from \code{update_admin_summaries}.}
}
\value{
A sorted varinfo tibble.
}
\description{
Organizes the varinfo table so that metadata is at the top, followed by
questions sorted by their most recent administration year and their
original order in the Qualtrics survey file (for the most recent admin).
}
\examples{
varinfo <- tibble::tibble(
  ITEM_NAME = c("AGE", "CONSENT"),
  ITEM_TYPE = c("question", "metadata"),
  MostRecentSurveyAdmin = c("2025", "2025"),
  QualtricsVariableName.2025 = c("Q2", "Q1")
)
current_meta <- tibble::tibble(QualtricsVariableName.2025 = c("Q1", "Q2"),
QualtricsVariableText.2025 = c("question one", "question two"))

sort_varinfo(joined_varinfo, current_meta,  c("2025", "2024"))
}
\seealso{
Other varinfo prep functions: 
\code{\link{add_catalog_display_flag}()},
\code{\link{add_dash_display_flag}()},
\code{\link{add_label_metadata}()},
\code{\link{check_and_export_unmatched}()},
\code{\link{get_catalog_varinfo}()},
\code{\link{get_dashboard_varinfo}()},
\code{\link{join_varinfo}()},
\code{\link{load_survey_column_info}()},
\code{\link{sync_column_schema}()},
\code{\link{update_survey_admin_summaries}()}
}
\concept{varinfo prep functions}
