% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepair.R
\name{st_prepair}
\alias{st_prepair}
\title{Automatic repair of single polygons}
\usage{
st_prepair(x)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}, it only works POLYGON or MULTIPOLYGON are supported}
}
\value{
an object of class \code{sf}, \code{sfc} or \code{sfg} depending on whether
the input is respectively \code{sf}, \code{sfc} or \code{sfg}
}
\description{
\code{st_prepair} repairs single polygons according to the international standards ISO 19107
using a constrained triangulation approach.
}
\details{
\code{st_prepair} supports two algorithms:
\itemize{
\item oddeven: an extension of the odd-even algorithm to handle GIS polygons containing inner rings and degeneracies;
\item setdiff: one where we follow a point set difference rule for the rings (outer - inner).
}
}
\examples{
\dontrun{
library(sf)
p1 <- st_as_sfc("POLYGON((0 0, 0 10, 10 0, 10 10, 0 0))")
st_is_valid(p1)
st_prepair(p1)
st_is_valid(st_prepair(p1))
}

}
\references{
Ledoux, H., Arroyo Ohori, K., and Meijers, M. (2014).
A triangulation-based approach to automatically repair GIS polygons.
Computers & Geosciences 66:121–131.

van Oosterom P., Quak W., Tijssen T. (2005).
About Invalid, Valid and Clean Polygons.
In: Developments in Spatial Data Handling. Springer, Berlin, Heidelberg
}
\seealso{
\code{sf::st_make_valid} for another approach to fix broken polygons
}
