% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepair.R
\name{st_prepair}
\alias{st_prepair}
\title{Polygon repair}
\usage{
st_prepair(x, min_area = 0, point_set = FALSE)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{min_area}{mininum area to keep in output}

\item{point_set}{if TRUE use the point set difference algorithm for the rings (outer-inner)
by default (point_set = FALSE) the odd-even algorithm is used}
}
\description{
Automatically repair single polygons (according to the OGC Simple Features / ISO19107 rules) using a constrained triangulation approach.
}
\examples{
library(sf)
p1 <- st_as_sfc("POLYGON((0 0, 0 10, 10 0, 10 10, 0 0))")
st_is_valid(p1)
st_prepair(p1)
st_is_valid(st_prepair(p1))
}
\seealso{
\code{sf::st_make_valid} for another approach at fixing broken polygons
}
