% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{st_probustness}
\alias{st_probustness}
\title{Compute polygons robustness}
\usage{
st_probustness(x)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}, it only works with POLYGON or MULTIPOLYGON
are supported}
}
\value{
a numeric vector
}
\description{
Compute polygons robustness, the higher the value the more robust the polygon is.
robustness can be defined as the largest  distance  of  moving  a  vertex,
while  the  result  is  still  a  valid  polygon
}
\examples{
library(sf)
p1 <- st_as_sfc("POLYGON((0 0, 0 10, 10 0, 10 10, 0 0))")
st_is_valid(p1)
st_probustness(p1)
st_probustness(st_prepair(p1))

}
\references{
Ledoux, H., Arroyo Ohori, K., and Meijers, M. (2014).
A triangulation-based approach to automatically repair GIS polygons.
Computers & Geosciences 66:121–131.

van Oosterom P., Quak W., Tijssen T. (2005).
About Invalid, Valid and Clean Polygons.
In: Developments in Spatial Data Handling. Springer, Berlin, Heidelberg
}
