% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_survey_admin_summaries.R
\name{update_survey_admin_summaries}
\alias{update_survey_admin_summaries}
\title{Update Survey Administration Summary Columns}
\usage{
update_survey_admin_summaries(df, recency_order = NULL)
}
\arguments{
\item{df}{The varinfo data frame.}

\item{recency_order}{Optional vector of admin IDs in order (most recent first).}
}
\description{
Looks for columns in the varinfo with the pattern QualtricsVariableName.XXX
where XXX is the survey administration year, semester, or other date id
(columns can also be supplied as a vector in recency order). If a given
variable (row in the varinfo) is not NA for a column, we assume that question
was asked in that survey administration. Generates two new columns:
\enumerate{
\item \code{AllSurveyAdmin} with a comma separated list of admin IDs and
\item \code{MostRecentSurveyAdmin} with only the most recent admin ID.
}
}
\examples{
test_df <- tibble::tibble(
  ITEM_NAME = c("Q1", "Q2"),
  QualtricsVariableName.2024 = c("Q1", NA),
  QualtricsVariableName.2025 = c("Q1", "Q2")
)

update_survey_admin_summaries(test_df, recency_order = c("2025", "2024"))
}
\seealso{
Other varinfo prep functions: 
\code{\link{add_catalog_display_flag}()},
\code{\link{add_dash_display_flag}()},
\code{\link{add_label_metadata}()},
\code{\link{check_and_export_unmatched}()},
\code{\link{get_catalog_varinfo}()},
\code{\link{get_dashboard_varinfo}()},
\code{\link{join_varinfo}()},
\code{\link{load_survey_column_info}()},
\code{\link{sort_varinfo}()},
\code{\link{sync_column_schema}()}
}
\concept{varinfo prep functions}
